refactor: actualizar la configuración de ESLint para desactivar reglas específicas y ajustar el puerto de inicio en package.json. Se corrigen etiquetas en el formulario de registro de productos y se eliminan datos de productos obsoletos. Además, se realizan ajustes en la gestión de inventario y se mejora la validación de ID en varios componentes, optimizando la experiencia del usuario.

+++ presentation/components/inventory/refill-dialog.tsx
@@ -35,7 +35,7 @@ export function RefillDialog({ product }: RefillDialogProps) {
     });
     try {
       toast.promise(
-        refillProduct(product.id, {
+        refillProduct(product.id as bigint, {
           refill: quantity,
           type: product.type,
         }),
